themes translate left

+++ app/Http/Controllers/SkillController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\SkillRequest;
 use App\Models\Skill;
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 class SkillController extends Controller
@@ -28,6 +29,7 @@ class SkillController extends Controller
         DB::beginTransaction();
         try {
             $skill = Skill::findOrFail($id);
+
             $skill->update($request->validated());
             DB::commit();
             return redirect()->back()->with('success', 'Skill updated successfully');
@@ -43,14 +45,34 @@ class SkillController extends Controller
         DB::beginTransaction();
         try {
             $skill = Skill::findOrFail($id);
-            if ($skill) {
+            if ($skill && $skill->user_id === auth()->id()) {
                 $skill->delete();
                 DB::commit();
-                return redirect()->back()->with('success', 'Project Deleted successfully');
+                return redirect()->back()->with('success', 'Skill deleted successfully.');
             }
+            auth()->user()->skills()->detach($skill->id);
+            DB::commit();
+            return redirect()->back()->with('success', 'Skill removed from your list.');
         } catch (Exception $e) {
             DB::rollback();
-            return redirect()->back()->with('failure', 'Project delete failed due to' . $e->getMessage());
+            return redirect()->back()->with('failure', 'Skills delete failed due to' . $e->getMessage());
+        }
+    }
+    public function copySkill($id)
+    {
+        DB::beginTransaction();
+        try {
+            $skill = Skill::findOrFail($id);
+            Skill::create([
+                'name' => $skill->name,
+                'level' => $skill->level,
+                'user_id' => Auth::id(),
+            ]);
+            DB::commit();
+            return redirect()->back()->with('success', 'Skill copied successfully');
+        } catch (Exception $e) {
+            DB::rollback();
+            return redirect()->back()->with('failure', 'Skill copy failed due to' . $e->getMessage());
         }
     }
 }
